fix(settings): govern the 外观 page structure, grouping, and copy
The 外观 page carried four leftovers from the moves that reshaped it —
identity going to 通用 (P2-13), the Astryx open-group rebuild (#1972), and
the settings-page kit (#1991).

- Page root: `AppearanceSettingsPage` wrapped a `ThemeSettingsPage` that
  opened a `SettingsPage` of its own, so the page grid nested inside itself
  (two `.settingsPageStack` in the DOM). The page IS the theme page, as its
  own comment said, so it is now one component with one root. Same defect on
  通用, where the identity block was a `*Page` rendered inside another page:
  it is a section of someone else's page and is now named, filed, and shaped
  as one (`personalization-settings-section.tsx`).
- Grouping: the 14 option tiles were loose in the accessibility tree — the
  sections were unnamed landmarks and the visible 编辑器主题 / 产品色调
  labels had no programmatic relationship to the grids they head. Sections
  now carry `titleId`, and each grid is a `role="group"` named by its label.
- Copy: the nav description still advertised 界面语言 on 外观, a page that
  has not had it since P2-13; 通用, which does, never mentioned it. Both
  descriptions now say what their page holds.
- Dead copy: `sections.paletteHelp` was defined in both locales and read by
  nobody, while the palette section reached into `appearance.persistenceHelp`
  for its lede. The palette section now takes its lede from `sections`, the
  same namespace the theme section above it uses, and the unread key is gone.

`AppearanceSettingsPage` also stops taking a `settings` prop it only passed
down one level to a component that never read it.

Contract: `astryx-component-behavior` pins one page stack, the four
labelledby anchors, three named groups, no 界面语言 on 外观, and that the
identity block opens no page stack of its own.

Verified: desktop typecheck; format:check; console/a11y/copy checks;
dead-css; 1754 desktop tests; and both pages re-rendered in a real Electron
fixture window (外观 unchanged apart from the two copy lines, 通用 pixel
identical).
