+++ apps/desktop/src/renderer/locales/settings-navigation-copy.ts
@@ -15,8 +15,8 @@ const SETTINGS_NAVIGATION_COPY_BY_LOCALE = {
       system: '系统',
     },
     sections: {
-      general: { label: '通用', description: '隐身、启动、对话默认与网络代理等系统偏好。' },
-      appearance: { label: '外观', description: '主题、配色与界面语言。' },
+      general: { label: '通用', description: '显示名称与界面语言、隐私与通知、对话默认与网络代理。' },
+      appearance: { label: '外观', description: '界面主题与调色板。' },
       projects: { label: '项目', description: '管理 Maka 认识的项目，以及新对话默认打开哪一个。' },
       models: { label: '模型', description: '模型连接、API key 与 OAuth 订阅管理。' },
       subagents: { label: '子 Agent', description: '配置主 Agent 可以自动选择的子 Agent、能力边界与模型。' },
@@ -39,8 +39,8 @@ const SETTINGS_NAVIGATION_COPY_BY_LOCALE = {
       system: 'System',
     },
     sections: {
-      general: { label: 'General', description: 'Privacy, startup, conversation defaults, and network proxy preferences.' },
-      appearance: { label: 'Appearance', description: 'Theme, color palette, and interface language.' },
+      general: { label: 'General', description: 'Display name and interface language, privacy and notifications, conversation defaults, and network proxy.' },
+      appearance: { label: 'Appearance', description: 'Interface theme and color palette.' },
       projects: { label: 'Projects', description: 'Manage the projects Maka knows about, and which one new conversations open in.' },
       models: { label: 'Models', description: 'Model connections, API keys, and OAuth subscriptions.' },
       subagents: { label: 'Subagents', description: 'Configure the subagents, capability boundaries, and models the main agent may select.' },
